+++ src/jest.config.cjs
@@ -2,12 +2,11 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  roots: ['<rootDir>/packages/core/src', '<rootDir>/packages/contracts/src'],
+  roots: ['<rootDir>/packages/core/test', '<rootDir>/packages/contracts/test'],
   testMatch: ['**/*.test.ts'],
   collectCoverageFrom: [
     'packages/core/src/**/*.ts',
     'packages/contracts/src/**/*.ts',
-    '!**/*.test.ts',
   ],
   coverageDirectory: 'coverage',
   coverageReporters: ['json-summary', 'text'],